# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Binance Futures Paper Trading Research Platform API
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
...
paths:
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get all paper trades
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed]
        - name: direction
          in: query
          schema:
            type: string
            enum: [LONG, SHORT]
        - name: result
          in: query
          schema:
            type: string
            enum: [WIN, LOSS, BREAKEVEN]
        - name: source
          in: query
          schema:
            $ref: "#/components/schemas/SignalSource"
        - name: scannerType
          in: query
          schema:
            $ref: "#/components/schemas/ScannerType"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Trades list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperTrade"
...
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get trade by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTrade"
        "404":
          description: Not found
  /trades/{id}/close:
    post:
      operationId: closeTrade
      tags: [trades]
      summary: Manually close a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseTradeInput"
      responses:
        "200":
          description: Trade closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTrade"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignalSource:
      type: string
      enum: [TECHNICAL, SMC]
    ScannerType:
      type: string
      enum: [TECHNICAL_SCANNER, SMC_SCANNER]
...
    DiagnosticDetails:
      type: object
      additionalProperties: true
...
    PaperTrade:
      type: object
      required: [id, tradeId, symbol, direction, entryPrice, stopLoss, tp1, tp2, tp3, quantity, signalScore, reason, status, openedAt]
      properties:
        id:
          type: integer
        tradeId:
          type: string
        symbol:
          type: string
        direction:
          type: string
          enum: [LONG, SHORT]
        source:
          $ref: "#/components/schemas/SignalSource"
        scannerType:
          $ref: "#/components/schemas/ScannerType"
        strategyType:
          type: string
        strategyLabel:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        smcScore:
          type: ["number", "null"]
        smcDetails:
          oneOf:
            - $ref: "#/components/schemas/DiagnosticDetails"
            - type: "null"
        entryPrice:
          type: number
        stopLoss:
          type: number
        currentSl:
          type: ["number", "null"]
        tp1:
          type: number
        tp2:
          type: number
        tp3:
          type: number
        quantity:
          type: number
        signalScore:
          type: number
        signalGrade:
          type: ["string", "null"]
        reason:
          type: string
        slReason:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed]
        result:
          type: ["string", "null"]
          enum: [WIN, LOSS, BREAKEVEN, null]
        tp1Hit:
          type: boolean
        tp2Hit:
          type: boolean
        tp3Hit:
          type: boolean
        exitPrice:
          type: ["number", "null"]
        exitReason:
          type: ["string", "null"]
        pnl:
          type: ["number", "null"]
        pnlPercent:
          type: ["number", "null"]
        holdingDurationMinutes:
          type: ["integer", "null"]
        openedAt:
          type: string
        closedAt:
          type: ["string", "null"]
    CloseTradeInput:
      type: object
      required: [exitPrice, exitReason]
      properties:
        exitPrice:
          type: number
        exitReason:
          type: string
...